app/requirements.txt
instagrapi>=2.0.0
ffmpeg-python>=0.2.0
Pillow>=10.0.0
imagehash>=4.3.1
numpy>=1.26.0
torch>=2.1.0
open_clip_torch>=2.24.0
requests>=2.31.0
